Relocate some long tests
