+++ server/jetstream_cluster_3_test.go
@@ -5790,132 +5790,6 @@ func TestJetStreamClusterDetectOrphanNRGs(t *testing.T) {
 	require_True(t, s.lookupRaftNode(ogn) == nil)
 }
 
-func TestJetStreamClusterRestartThenScaleStreamReplicas(t *testing.T) {
-	t.Skip("This test takes too long, need to make shorter")
-
-	c := createJetStreamClusterExplicit(t, "R3S", 3)
-	defer c.shutdown()
-
-	s := c.randomNonLeader()
-	nc, js := jsClientConnect(t, s)
-	defer nc.Close()
-
-	nc2, producer := jsClientConnect(t, s)
-	defer nc2.Close()
-
-	_, err := js.AddStream(&nats.StreamConfig{
-		Name:     "TEST",
-		Subjects: []string{"foo"},
-		Replicas: 3,
-	})
-	require_NoError(t, err)
-	c.waitOnStreamLeader(globalAccountName, "TEST")
-
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	end := time.Now().Add(2 * time.Second)
-	for time.Now().Before(end) {
-		producer.Publish("foo", []byte(strings.Repeat("A", 128)))
-		time.Sleep(time.Millisecond)
-	}
-
-	var wg sync.WaitGroup
-	for i := 0; i < 5; i++ {
-		sub, err := js.PullSubscribe("foo", fmt.Sprintf("C-%d", i))
-		require_NoError(t, err)
-
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			for range time.NewTicker(10 * time.Millisecond).C {
-				select {
-				case <-ctx.Done():
-					return
-				default:
-				}
-
-				msgs, err := sub.Fetch(1)
-				if err != nil && !errors.Is(err, nats.ErrTimeout) && !errors.Is(err, nats.ErrConnectionClosed) {
-					t.Logf("Pull Error: %v", err)
-				}
-				for _, msg := range msgs {
-					msg.Ack()
-				}
-			}
-		}()
-	}
-	c.lameDuckRestartAll()
-	c.waitOnStreamLeader(globalAccountName, "TEST")
-
-	// Swap the logger to try to detect the condition after the restart.
-	loggers := make([]*captureDebugLogger, 3)
-	for i, srv := range c.servers {
-		l := &captureDebugLogger{dbgCh: make(chan string, 10)}
-		loggers[i] = l
-		srv.SetLogger(l, true, false)
-	}
-	condition := `Direct proposal ignored, not leader (state: CLOSED)`
-	errCh := make(chan error, 10)
-
-	wg.Add(1)
-	go func() {
-		defer wg.Done()
-		for {
-			select {
-			case dl := <-loggers[0].dbgCh:
-				if strings.Contains(dl, condition) {
-					errCh <- errors.New(condition)
-				}
-			case dl := <-loggers[1].dbgCh:
-				if strings.Contains(dl, condition) {
-					errCh <- errors.New(condition)
-				}
-			case dl := <-loggers[2].dbgCh:
-				if strings.Contains(dl, condition) {
-					errCh <- errors.New(condition)
-				}
-			case <-ctx.Done():
-				return
-			}
-		}
-	}()
-
-	// Start publishing again for a while.
-	end = time.Now().Add(2 * time.Second)
-	for time.Now().Before(end) {
-		producer.Publish("foo", []byte(strings.Repeat("A", 128)))
-		time.Sleep(time.Millisecond)
-	}
-
-	// Try to do a stream edit back to R=1 after doing all the upgrade.
-	info, _ := js.StreamInfo("TEST")
-	sconfig := info.Config
-	sconfig.Replicas = 1
-	_, err = js.UpdateStream(&sconfig)
-	require_NoError(t, err)
-
-	// Leave running for some time after the update.
-	time.Sleep(2 * time.Second)
-
-	info, _ = js.StreamInfo("TEST")
-	sconfig = info.Config
-	sconfig.Replicas = 3
-	_, err = js.UpdateStream(&sconfig)
-	require_NoError(t, err)
-
-	select {
-	case e := <-errCh:
-		t.Fatalf("Bad condition on raft node: %v", e)
-	case <-time.After(2 * time.Second):
-		// Done
-	}
-
-	// Stop goroutines and wait for them to exit.
-	cancel()
-	wg.Wait()
-}
-
 // https://github.com/nats-io/nats-server/issues/4732
 func TestJetStreamClusterStreamLimitsOnScaleUpAndMove(t *testing.T) {
 	tmpl := `
