+++ server/jetstream_cluster_4_test.go
@@ -25,7 +25,6 @@ import (
 	"os"
 	"path/filepath"
 	"runtime"
-	"slices"
 	"strconv"
 	"strings"
 	"sync"
@@ -1680,573 +1679,6 @@ NextServer:
 	}
 }
 
-func TestJetStreamClusterBusyStreams(t *testing.T) {
-	t.Skip("Too long for CI at the moment")
-	type streamSetup struct {
-		config    *nats.StreamConfig
-		consumers []*nats.ConsumerConfig
-		subjects  []string
-	}
-	type job func(t *testing.T, nc *nats.Conn, js nats.JetStreamContext, c *cluster)
-	type testParams struct {
-		cluster         string
-		streams         []*streamSetup
-		producers       int
-		consumers       int
-		restartAny      bool
-		restartWait     time.Duration
-		ldmRestart      bool
-		rolloutRestart  bool
-		restarts        int
-		checkHealthz    bool
-		jobs            []job
-		expect          job
-		duration        time.Duration
-		producerMsgs    int
-		producerMsgSize int
-	}
-	test := func(t *testing.T, test *testParams) {
-		conf := `
-                listen: 127.0.0.1:-1
-                http: 127.0.0.1:-1
-                server_name: %s
-                jetstream: {
-                        domain: "cloud"
-                        store_dir: '%s',
-                }
-                cluster {
-                        name: %s
-                        listen: 127.0.0.1:%d
-                        routes = [%s]
-                }
-                server_tags: ["test"]
-                system_account: sys
-
-                no_auth_user: js
-                accounts {
-                        sys { users = [ { user: sys, pass: sys } ] }
-
-                        js  { jetstream = enabled
-                              users = [ { user: js, pass: js } ]
-                        }
-                }`
-		c := createJetStreamClusterWithTemplate(t, conf, test.cluster, 3)
-		defer c.shutdown()
-		for _, s := range c.servers {
-			s.optsMu.Lock()
-			s.opts.LameDuckDuration = 15 * time.Second
-			s.opts.LameDuckGracePeriod = -15 * time.Second
-			s.optsMu.Unlock()
-		}
-
-		nc, js := jsClientConnect(t, c.randomServer())
-		defer nc.Close()
-
-		var wg sync.WaitGroup
-		for _, stream := range test.streams {
-			stream := stream
-			wg.Add(1)
-			go func() {
-				defer wg.Done()
-				_, err := js.AddStream(stream.config)
-				require_NoError(t, err)
-
-				for _, consumer := range stream.consumers {
-					_, err := js.AddConsumer(stream.config.Name, consumer)
-					require_NoError(t, err)
-				}
-			}()
-		}
-		wg.Wait()
-
-		ctx, cancel := context.WithTimeout(context.Background(), test.duration)
-		defer cancel()
-		for _, stream := range test.streams {
-			payload := []byte(strings.Repeat("A", test.producerMsgSize))
-			stream := stream
-			subjects := stream.subjects
-
-			// Create publishers on different connections that sends messages
-			// to all the consumers subjects.
-			var n atomic.Uint64
-			for i := 0; i < test.producers; i++ {
-				wg.Add(1)
-				go func() {
-					defer wg.Done()
-					nc, js := jsClientConnect(t, c.randomServer())
-					defer nc.Close()
-
-					for range time.NewTicker(1 * time.Millisecond).C {
-						select {
-						case <-ctx.Done():
-							return
-						default:
-						}
-
-						for _, subject := range subjects {
-							msgID := nats.MsgId(fmt.Sprintf("n:%d", n.Load()))
-							_, err := js.Publish(subject, payload, nats.AckWait(200*time.Millisecond), msgID)
-							if err == nil {
-								if nn := n.Add(1); int(nn) >= test.producerMsgs {
-									return
-								}
-							}
-						}
-					}
-				}()
-			}
-
-			// Create multiple parallel pull subscribers per consumer config.
-			for i := 0; i < test.consumers; i++ {
-				for _, consumer := range stream.consumers {
-					wg.Add(1)
-
-					consumer := consumer
-					go func() {
-						defer wg.Done()
-
-						for attempts := 0; attempts < 60; attempts++ {
-							_, err := js.ConsumerInfo(stream.config.Name, consumer.Name)
-							if err != nil {
-								t.Logf("WRN: Failed creating pull subscriber: %v - %v - %v - %v",
-									consumer.FilterSubject, stream.config.Name, consumer.Name, err)
-							}
-						}
-						sub, err := js.PullSubscribe(consumer.FilterSubject, "", nats.Bind(stream.config.Name, consumer.Name))
-						if err != nil {
-							t.Logf("WRN: Failed creating pull subscriber: %v - %v - %v - %v",
-								consumer.FilterSubject, stream.config.Name, consumer.Name, err)
-							return
-						}
-						require_NoError(t, err)
-
-						for range time.NewTicker(100 * time.Millisecond).C {
-							select {
-							case <-ctx.Done():
-								return
-							default:
-							}
-
-							msgs, err := sub.Fetch(1, nats.MaxWait(200*time.Millisecond))
-							if err != nil {
-								continue
-							}
-							for _, msg := range msgs {
-								msg.Ack()
-							}
-
-							msgs, err = sub.Fetch(100, nats.MaxWait(200*time.Millisecond))
-							if err != nil {
-								continue
-							}
-							for _, msg := range msgs {
-								msg.Ack()
-							}
-						}
-					}()
-				}
-			}
-		}
-
-		for _, job := range test.jobs {
-			go job(t, nc, js, c)
-		}
-		if test.restarts > 0 {
-			wg.Add(1)
-			time.AfterFunc(test.restartWait, func() {
-				defer wg.Done()
-				for i := 0; i < test.restarts; i++ {
-					switch {
-					case test.restartAny:
-						s := c.servers[rand.Intn(len(c.servers))]
-						if test.ldmRestart {
-							s.lameDuckMode()
-						} else {
-							s.Shutdown()
-						}
-						s.WaitForShutdown()
-						c.restartServer(s)
-					case test.rolloutRestart:
-						for _, s := range c.servers {
-							if test.ldmRestart {
-								s.lameDuckMode()
-							} else {
-								s.Shutdown()
-							}
-							s.WaitForShutdown()
-							s = c.restartServer(s)
-
-							if test.checkHealthz {
-								hctx, hcancel := context.WithTimeout(ctx, 15*time.Second)
-								defer hcancel()
-
-							Healthz:
-								for range time.NewTicker(2 * time.Second).C {
-									select {
-									case <-hctx.Done():
-										break Healthz
-									default:
-									}
-
-									status := s.healthz(nil)
-									if status.StatusCode == 200 {
-										break Healthz
-									}
-								}
-							}
-						}
-					}
-					c.waitOnClusterReady()
-				}
-			})
-		}
-		test.expect(t, nc, js, c)
-		cancel()
-		wg.Wait()
-	}
-	stepDown := func(nc *nats.Conn, streamName string) {
-		nc.Request(fmt.Sprintf(JSApiStreamLeaderStepDownT, streamName), nil, time.Second)
-	}
-	getStreamDetails := func(t *testing.T, c *cluster, accountName, streamName string) *StreamDetail {
-		t.Helper()
-		srv := c.streamLeader(accountName, streamName)
-		jsz, err := srv.Jsz(&JSzOptions{Accounts: true, Streams: true, Consumer: true})
-		require_NoError(t, err)
-		for _, acc := range jsz.AccountDetails {
-			if acc.Name == accountName {
-				for _, stream := range acc.Streams {
-					if stream.Name == streamName {
-						return &stream
-					}
-				}
-			}
-		}
-		t.Error("Could not find account details")
-		return nil
-	}
-	checkMsgsEqual := func(t *testing.T, c *cluster, accountName, streamName string) {
-		state := getStreamDetails(t, c, accountName, streamName).State
-		var msets []*stream
-		for _, s := range c.servers {
-			acc, err := s.LookupAccount(accountName)
-			require_NoError(t, err)
-			mset, err := acc.lookupStream(streamName)
-			require_NoError(t, err)
-			msets = append(msets, mset)
-		}
-		for seq := state.FirstSeq; seq <= state.LastSeq; seq++ {
-			var msgId string
-			var smv StoreMsg
-			for _, mset := range msets {
-				mset.mu.RLock()
-				sm, err := mset.store.LoadMsg(seq, &smv)
-				mset.mu.RUnlock()
-				require_NoError(t, err)
-				if msgId == _EMPTY_ {
-					msgId = string(sm.hdr)
-				} else if msgId != string(sm.hdr) {
-					t.Fatalf("MsgIds do not match for seq %d: %q vs %q", seq, msgId, sm.hdr)
-				}
-			}
-		}
-	}
-	checkConsumer := func(t *testing.T, c *cluster, accountName, streamName, consumerName string) {
-		t.Helper()
-		var leader string
-		for _, s := range c.servers {
-			jsz, err := s.Jsz(&JSzOptions{Accounts: true, Streams: true, Consumer: true})
-			require_NoError(t, err)
-			for _, acc := range jsz.AccountDetails {
-				if acc.Name == accountName {
-					for _, stream := range acc.Streams {
-						if stream.Name == streamName {
-							for _, consumer := range stream.Consumer {
-								if leader == "" {
-									leader = consumer.Cluster.Leader
-								} else if leader != consumer.Cluster.Leader {
-									t.Errorf("There are two leaders for %s/%s: %s vs %s",
-										stream.Name, consumer.Name, leader, consumer.Cluster.Leader)
-								}
-							}
-						}
-					}
-				}
-			}
-		}
-	}
-
-	t.Run("R1F/rescale/R3F/sources:10/limits", func(t *testing.T) {
-		testDuration := 3 * time.Minute
-		totalStreams := 10
-		streams := make([]*streamSetup, totalStreams)
-		sources := make([]*nats.StreamSource, totalStreams)
-		for i := 0; i < totalStreams; i++ {
-			name := fmt.Sprintf("test:%d", i)
-			st := &streamSetup{
-				config: &nats.StreamConfig{
-					Name:      name,
-					Subjects:  []string{fmt.Sprintf("test.%d.*", i)},
-					Replicas:  1,
-					Retention: nats.LimitsPolicy,
-				},
-			}
-			st.subjects = append(st.subjects, fmt.Sprintf("test.%d.0", i))
-			sources[i] = &nats.StreamSource{Name: name}
-			streams[i] = st
-		}
-
-		// Create Source consumer.
-		sourceSetup := &streamSetup{
-			config: &nats.StreamConfig{
-				Name:      "source-test",
-				Replicas:  1,
-				Retention: nats.LimitsPolicy,
-				Sources:   sources,
-			},
-			consumers: make([]*nats.ConsumerConfig, 0),
-		}
-		cc := &nats.ConsumerConfig{
-			Name:          "A",
-			Durable:       "A",
-			FilterSubject: "test.>",
-			AckPolicy:     nats.AckExplicitPolicy,
-		}
-		sourceSetup.consumers = append(sourceSetup.consumers, cc)
-		streams = append(streams, sourceSetup)
-
-		scale := func(replicas int, wait time.Duration) job {
-			return func(t *testing.T, nc *nats.Conn, js nats.JetStreamContext, c *cluster) {
-				config := sourceSetup.config
-				time.AfterFunc(wait, func() {
-					config.Replicas = replicas
-					for i := 0; i < 10; i++ {
-						_, err := js.UpdateStream(config)
-						if err == nil {
-							return
-						}
-						time.Sleep(1 * time.Second)
-					}
-				})
-			}
-		}
-
-		expect := func(t *testing.T, nc *nats.Conn, js nats.JetStreamContext, c *cluster) {
-			// The source stream should not be stuck or be different from the other streams.
-			time.Sleep(testDuration + 1*time.Minute)
-			accName := "js"
-			streamName := "source-test"
-
-			// Check a few times to see if there are no changes in the number of messages.
-			var changed bool
-			var prevMsgs uint64
-			for i := 0; i < 10; i++ {
-				sinfo, err := js.StreamInfo(streamName)
-				if err != nil {
-					t.Logf("Error: %v", err)
-					time.Sleep(2 * time.Second)
-					continue
-				}
-				prevMsgs = sinfo.State.Msgs
-			}
-			for i := 0; i < 10; i++ {
-				sinfo, err := js.StreamInfo(streamName)
-				if err != nil {
-					t.Logf("Error: %v", err)
-					time.Sleep(2 * time.Second)
-					continue
-				}
-				changed = prevMsgs != sinfo.State.Msgs
-				prevMsgs = sinfo.State.Msgs
-				time.Sleep(2 * time.Second)
-			}
-			if !changed {
-				// Doing a leader step down should not cause the messages to change.
-				stepDown(nc, streamName)
-
-				for i := 0; i < 10; i++ {
-					sinfo, err := js.StreamInfo(streamName)
-					if err != nil {
-						t.Logf("Error: %v", err)
-						time.Sleep(2 * time.Second)
-						continue
-					}
-					changed = prevMsgs != sinfo.State.Msgs
-					prevMsgs = sinfo.State.Msgs
-					time.Sleep(2 * time.Second)
-				}
-				if changed {
-					t.Error("Stream msgs changed after the step down")
-				}
-			}
-
-			/////////////////////////////////////////////////////////////////////////////////////////
-			//                                                                                     //
-			//  The number of messages sourced should match the count from all the other streams.  //
-			//                                                                                     //
-			/////////////////////////////////////////////////////////////////////////////////////////
-			var expectedMsgs uint64
-			for i := 0; i < totalStreams; i++ {
-				name := fmt.Sprintf("test:%d", i)
-				sinfo, err := js.StreamInfo(name)
-				require_NoError(t, err)
-				expectedMsgs += sinfo.State.Msgs
-			}
-			sinfo, err := js.StreamInfo(streamName)
-			require_NoError(t, err)
-
-			gotMsgs := sinfo.State.Msgs
-			if gotMsgs != expectedMsgs {
-				t.Errorf("stream with sources has %v messages, but total sourced messages should be %v", gotMsgs, expectedMsgs)
-			}
-			checkConsumer(t, c, accName, streamName, "A")
-			checkMsgsEqual(t, c, accName, streamName)
-		}
-		test(t, &testParams{
-			cluster:        t.Name(),
-			streams:        streams,
-			producers:      10,
-			consumers:      10,
-			restarts:       1,
-			rolloutRestart: true,
-			ldmRestart:     true,
-			checkHealthz:   true,
-			// TODO(dlc) - If this overlaps with the scale jobs this test will fail.
-			// Leaders will be elected with partial state.
-			restartWait: 65 * time.Second,
-			jobs: []job{
-				scale(3, 15*time.Second),
-				scale(1, 30*time.Second),
-				scale(3, 60*time.Second),
-			},
-			expect:          expect,
-			duration:        testDuration,
-			producerMsgSize: 1024,
-			producerMsgs:    100_000,
-		})
-	})
-
-	t.Run("rollouts", func(t *testing.T) {
-		shared := func(t *testing.T, sc *nats.StreamConfig, tp *testParams) func(t *testing.T) {
-			return func(t *testing.T) {
-				testDuration := 3 * time.Minute
-				totalStreams := 30
-				consumersPerStream := 5
-				streams := make([]*streamSetup, totalStreams)
-				for i := 0; i < totalStreams; i++ {
-					name := fmt.Sprintf("test:%d", i)
-					st := &streamSetup{
-						config: &nats.StreamConfig{
-							Name:      name,
-							Subjects:  []string{fmt.Sprintf("test.%d.*", i)},
-							Replicas:  3,
-							Discard:   sc.Discard,
-							Retention: sc.Retention,
-							Storage:   sc.Storage,
-							MaxMsgs:   sc.MaxMsgs,
-							MaxBytes:  sc.MaxBytes,
-							MaxAge:    sc.MaxAge,
-						},
-						consumers: make([]*nats.ConsumerConfig, 0),
-					}
-					for j := 0; j < consumersPerStream; j++ {
-						subject := fmt.Sprintf("test.%d.%d", i, j)
-						name := fmt.Sprintf("A:%d:%d", i, j)
-						cc := &nats.ConsumerConfig{
-							Name:          name,
-							Durable:       name,
-							FilterSubject: subject,
-							AckPolicy:     nats.AckExplicitPolicy,
-						}
-						st.consumers = append(st.consumers, cc)
-						st.subjects = append(st.subjects, subject)
-					}
-					streams[i] = st
-				}
-				expect := func(t *testing.T, nc *nats.Conn, js nats.JetStreamContext, c *cluster) {
-					time.Sleep(testDuration + 1*time.Minute)
-					accName := "js"
-					for i := 0; i < totalStreams; i++ {
-						streamName := fmt.Sprintf("test:%d", i)
-						checkMsgsEqual(t, c, accName, streamName)
-					}
-				}
-				test(t, &testParams{
-					cluster:         t.Name(),
-					streams:         streams,
-					producers:       10,
-					consumers:       10,
-					restarts:        tp.restarts,
-					rolloutRestart:  tp.rolloutRestart,
-					ldmRestart:      tp.ldmRestart,
-					checkHealthz:    tp.checkHealthz,
-					restartWait:     tp.restartWait,
-					expect:          expect,
-					duration:        testDuration,
-					producerMsgSize: 1024,
-					producerMsgs:    100_000,
-				})
-			}
-		}
-		for prefix, st := range map[string]nats.StorageType{"R3F": nats.FileStorage, "R3M": nats.MemoryStorage} {
-			t.Run(prefix, func(t *testing.T) {
-				for rolloutType, params := range map[string]*testParams{
-					// Rollouts using graceful restarts and checking healthz.
-					"ldm": {
-						restarts:       1,
-						rolloutRestart: true,
-						ldmRestart:     true,
-						checkHealthz:   true,
-						restartWait:    45 * time.Second,
-					},
-					// Non graceful restarts calling Shutdown, but using healthz on startup.
-					"term": {
-						restarts:       1,
-						rolloutRestart: true,
-						ldmRestart:     false,
-						checkHealthz:   true,
-						restartWait:    45 * time.Second,
-					},
-				} {
-					t.Run(rolloutType, func(t *testing.T) {
-						t.Run("limits", shared(t, &nats.StreamConfig{
-							Retention: nats.LimitsPolicy,
-							Storage:   st,
-						}, params))
-						t.Run("wq", shared(t, &nats.StreamConfig{
-							Retention: nats.WorkQueuePolicy,
-							Storage:   st,
-						}, params))
-						t.Run("interest", shared(t, &nats.StreamConfig{
-							Retention: nats.InterestPolicy,
-							Storage:   st,
-						}, params))
-						t.Run("limits:dn:max-per-subject", shared(t, &nats.StreamConfig{
-							Retention:         nats.LimitsPolicy,
-							Storage:           st,
-							MaxMsgsPerSubject: 1,
-							Discard:           nats.DiscardNew,
-						}, params))
-						t.Run("wq:dn:max-msgs", shared(t, &nats.StreamConfig{
-							Retention: nats.WorkQueuePolicy,
-							Storage:   st,
-							MaxMsgs:   10_000,
-							Discard:   nats.DiscardNew,
-						}, params))
-						t.Run("wq:dn-per-subject:max-msgs", shared(t, &nats.StreamConfig{
-							Retention:            nats.WorkQueuePolicy,
-							Storage:              st,
-							MaxMsgs:              10_000,
-							MaxMsgsPerSubject:    100,
-							Discard:              nats.DiscardNew,
-							DiscardNewPerSubject: true,
-						}, params))
-					})
-				}
-			})
-		}
-	})
-}
-
 // https://github.com/nats-io/nats-server/issues/5488
 func TestJetStreamClusterSingleMaxConsumerUpdate(t *testing.T) {
 	c := createJetStreamClusterExplicit(t, "R3S", 3)
@@ -2876,367 +2308,6 @@ func TestJetStreamClusterKeyValueDesyncAfterHardKill(t *testing.T) {
 
 	err = checkState(t, c, "$G", "KV_inconsistency")
 	require_NoError(t, err)
-}
-
-func TestJetStreamClusterKeyValueSync(t *testing.T) {
-	t.Skip("Too long for CI at the moment")
-	c := createJetStreamClusterExplicit(t, "R3S", 3)
-	defer c.shutdown()
-
-	for _, s := range c.servers {
-		s.optsMu.Lock()
-		s.opts.LameDuckDuration = 15 * time.Second
-		s.opts.LameDuckGracePeriod = -15 * time.Second
-		s.optsMu.Unlock()
-	}
-	s := c.randomNonLeader()
-	connect := func(t *testing.T) (*nats.Conn, nats.JetStreamContext) {
-		return jsClientConnect(t, s)
-	}
-
-	const accountName = "$G"
-	const letterBytes = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
-	createData := func(n int) []byte {
-		b := make([]byte, n)
-		for i := range b {
-			b[i] = letterBytes[rand.Intn(len(letterBytes))]
-		}
-		return b
-	}
-	getOrCreateKvStore := func(kvname string) (nats.KeyValue, error) {
-		_, js := connect(t)
-		kvExists := false
-		existingKvnames := js.KeyValueStoreNames()
-		for existingKvname := range existingKvnames {
-			if existingKvname == kvname {
-				kvExists = true
-				break
-			}
-		}
-		if !kvExists {
-			return js.CreateKeyValue(&nats.KeyValueConfig{
-				Bucket:   kvname,
-				Replicas: 3,
-				Storage:  nats.FileStorage,
-			})
-		} else {
-			return js.KeyValue(kvname)
-		}
-	}
-	abs := func(x int64) int64 {
-		if x < 0 {
-			return -x
-		}
-		return x
-	}
-	var counter int64
-	var errorCounter int64
-
-	getStreamDetails := func(t *testing.T, c *cluster, accountName, streamName string) *StreamDetail {
-		t.Helper()
-		srv := c.streamLeader(accountName, streamName)
-		if srv == nil {
-			return nil
-		}
-		jsz, err := srv.Jsz(&JSzOptions{Accounts: true, Streams: true, Consumer: true})
-		require_NoError(t, err)
-		for _, acc := range jsz.AccountDetails {
-			if acc.Name == accountName {
-				for _, stream := range acc.Streams {
-					if stream.Name == streamName {
-						return &stream
-					}
-				}
-			}
-		}
-		t.Error("Could not find account details")
-		return nil
-	}
-	checkState := func(t *testing.T, c *cluster, accountName, streamName string) error {
-		t.Helper()
-
-		leaderSrv := c.streamLeader(accountName, streamName)
-		if leaderSrv == nil {
-			return fmt.Errorf("no leader server found for stream %q", streamName)
-		}
-		streamLeader := getStreamDetails(t, c, accountName, streamName)
-		if streamLeader == nil {
-			return fmt.Errorf("no leader found for stream %q", streamName)
-		}
-		var errs []error
-		for _, srv := range c.servers {
-			if srv == leaderSrv {
-				// Skip self
-				continue
-			}
-			acc, err := srv.LookupAccount(accountName)
-			require_NoError(t, err)
-			stream, err := acc.lookupStream(streamName)
-			require_NoError(t, err)
-			state := stream.state()
-
-			if state.Msgs != streamLeader.State.Msgs {
-				err := fmt.Errorf("[%s] Leader %v has %d messages, Follower %v has %d messages",
-					streamName, leaderSrv, streamLeader.State.Msgs,
-					srv, state.Msgs,
-				)
-				errs = append(errs, err)
-			}
-			if state.FirstSeq != streamLeader.State.FirstSeq {
-				err := fmt.Errorf("[%s] Leader %v FirstSeq is %d, Follower %v is at %d",
-					streamName, leaderSrv, streamLeader.State.FirstSeq,
-					srv, state.FirstSeq,
-				)
-				errs = append(errs, err)
-			}
-			if state.LastSeq != streamLeader.State.LastSeq {
-				err := fmt.Errorf("[%s] Leader %v LastSeq is %d, Follower %v is at %d",
-					streamName, leaderSrv, streamLeader.State.LastSeq,
-					srv, state.LastSeq,
-				)
-				errs = append(errs, err)
-			}
-			if state.NumDeleted != streamLeader.State.NumDeleted {
-				err := fmt.Errorf("[%s] Leader %v NumDeleted is %d, Follower %v is at %d\nSTATE_A: %+v\nSTATE_B: %+v\n",
-					streamName, leaderSrv, streamLeader.State.NumDeleted,
-					srv, state.NumDeleted, streamLeader.State, state,
-				)
-				errs = append(errs, err)
-			}
-		}
-		if len(errs) > 0 {
-			return errors.Join(errs...)
-		}
-		return nil
-	}
-
-	checkMsgsEqual := func(t *testing.T, accountName, streamName string) error {
-		// Gather all the streams replicas and compare contents.
-		msets := make(map[*Server]*stream)
-		for _, s := range c.servers {
-			acc, err := s.LookupAccount(accountName)
-			if err != nil {
-				return err
-			}
-			mset, err := acc.lookupStream(streamName)
-			if err != nil {
-				return err
-			}
-			msets[s] = mset
-		}
-
-		str := getStreamDetails(t, c, accountName, streamName)
-		if str == nil {
-			return fmt.Errorf("could not get stream leader state")
-		}
-		state := str.State
-		for seq := state.FirstSeq; seq <= state.LastSeq; seq++ {
-			var msgId string
-			var smv StoreMsg
-			for replica, mset := range msets {
-				mset.mu.RLock()
-				sm, err := mset.store.LoadMsg(seq, &smv)
-				mset.mu.RUnlock()
-				if err != nil {
-					if err == ErrStoreMsgNotFound || err == errDeletedMsg {
-						// Skip these.
-					} else {
-						t.Logf("WRN: Error loading message (seq=%d) from stream %q on replica %q: %v", seq, streamName, replica, err)
-					}
-					continue
-				}
-				if msgId == _EMPTY_ {
-					msgId = string(sm.hdr)
-				} else if msgId != string(sm.hdr) {
-					t.Errorf("MsgIds do not match for seq %d on stream %q: %q vs %q", seq, streamName, msgId, sm.hdr)
-				}
-			}
-		}
-		return nil
-	}
-
-	keyUpdater := func(ctx context.Context, cancel context.CancelFunc, kvname string, numKeys int) {
-		kv, err := getOrCreateKvStore(kvname)
-		if err != nil {
-			t.Fatalf("[%s]:%v", kvname, err)
-		}
-		for i := 0; i < numKeys; i++ {
-			key := fmt.Sprintf("key-%d", i)
-			kv.Create(key, createData(160))
-		}
-		lastData := make(map[string][]byte)
-		revisions := make(map[string]uint64)
-		for {
-			select {
-			case <-ctx.Done():
-				return
-			default:
-			}
-			r := rand.Intn(numKeys)
-			key := fmt.Sprintf("key-%d", r)
-
-			for i := 0; i < 5; i++ {
-				_, err := kv.Get(key)
-				if err != nil {
-					atomic.AddInt64(&errorCounter, 1)
-					if err == nats.ErrKeyNotFound {
-						t.Logf("WRN: Key not found! [%s/%s] - [%s]", kvname, key, err)
-						cancel()
-					}
-				}
-			}
-
-			k, err := kv.Get(key)
-			if err != nil {
-				atomic.AddInt64(&errorCounter, 1)
-			} else {
-				if revisions[key] != 0 && abs(int64(k.Revision())-int64(revisions[key])) < 2 {
-					lastDataVal, ok := lastData[key]
-					if ok && k.Revision() == revisions[key] && slices.Compare(lastDataVal, k.Value()) != 0 {
-						t.Logf("data loss [%s/%s][rev:%d] expected:[%v] is:[%v]", kvname, key, revisions[key], string(lastDataVal), string(k.Value()))
-					}
-				}
-				newData := createData(160)
-				revisions[key], err = kv.Update(key, newData, k.Revision())
-				if err != nil && err != nats.ErrTimeout {
-					atomic.AddInt64(&errorCounter, 1)
-				} else {
-					lastData[key] = newData
-				}
-				atomic.AddInt64(&counter, 1)
-			}
-		}
-	}
-
-	streamCount := 50
-	keysCount := 100
-	streamPrefix := "IKV"
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
-	defer cancel()
-
-	// The keyUpdaters will run for less time.
-	kctx, kcancel := context.WithTimeout(context.Background(), 3*time.Minute)
-	defer kcancel()
-
-	var wg sync.WaitGroup
-	var streams []string
-	for i := 0; i < streamCount; i++ {
-		streamName := fmt.Sprintf("%s-%d", streamPrefix, i)
-		streams = append(streams, "KV_"+streamName)
-
-		wg.Add(1)
-		go func(i int) {
-			defer wg.Done()
-			keyUpdater(kctx, cancel, streamName, keysCount)
-		}(i)
-	}
-
-	debug := false
-	nc2, _ := jsClientConnect(t, s)
-	if debug {
-		go func() {
-			for range time.NewTicker(5 * time.Second).C {
-				select {
-				case <-ctx.Done():
-					return
-				default:
-				}
-				for _, str := range streams {
-					leaderSrv := c.streamLeader(accountName, str)
-					if leaderSrv == nil {
-						continue
-					}
-					streamLeader := getStreamDetails(t, c, accountName, str)
-					if streamLeader == nil {
-						continue
-					}
-					t.Logf("|------------------------------------------------------------------------------------------------------------------------|")
-					lstate := streamLeader.State
-					t.Logf("| %-10s | %-10s | msgs:%-10d | bytes:%-10d | deleted:%-10d | first:%-10d | last:%-10d |",
-						str, leaderSrv.String()+"*", lstate.Msgs, lstate.Bytes, lstate.NumDeleted, lstate.FirstSeq, lstate.LastSeq,
-					)
-					for _, srv := range c.servers {
-						if srv == leaderSrv {
-							continue
-						}
-						acc, err := srv.LookupAccount(accountName)
-						if err != nil {
-							continue
-						}
-						stream, err := acc.lookupStream(str)
-						if err != nil {
-							t.Logf("Error looking up stream %s on %s replica", str, srv)
-							continue
-						}
-						state := stream.state()
-
-						unsynced := lstate.Msgs != state.Msgs || lstate.Bytes != state.Bytes ||
-							lstate.NumDeleted != state.NumDeleted || lstate.FirstSeq != state.FirstSeq || lstate.LastSeq != state.LastSeq
-
-						var result string
-						if unsynced {
-							result = "UNSYNCED"
-						}
-						t.Logf("| %-10s | %-10s | msgs:%-10d | bytes:%-10d | deleted:%-10d | first:%-10d | last:%-10d | %s",
-							str, srv, state.Msgs, state.Bytes, state.NumDeleted, state.FirstSeq, state.LastSeq, result,
-						)
-					}
-				}
-				t.Logf("|------------------------------------------------------------------------------------------------------------------------| %v", nc2.ConnectedUrl())
-			}
-		}()
-	}
-
-	checkStreams := func(t *testing.T) {
-		for _, str := range streams {
-			checkFor(t, time.Minute, 500*time.Millisecond, func() error {
-				return checkState(t, c, accountName, str)
-			})
-			checkFor(t, time.Minute, 500*time.Millisecond, func() error {
-				return checkMsgsEqual(t, accountName, str)
-			})
-		}
-	}
-
-Loop:
-	for range time.NewTicker(30 * time.Second).C {
-		select {
-		case <-ctx.Done():
-			break Loop
-		default:
-		}
-		rollout := func(t *testing.T) {
-			for _, s := range c.servers {
-				// For graceful mode
-				s.lameDuckMode()
-				s.WaitForShutdown()
-				s = c.restartServer(s)
-
-				hctx, hcancel := context.WithTimeout(context.Background(), 15*time.Second)
-				defer hcancel()
-
-			Healthz:
-				for range time.NewTicker(2 * time.Second).C {
-					select {
-					case <-hctx.Done():
-					default:
-					}
-
-					status := s.healthz(nil)
-					if status.StatusCode == 200 {
-						break Healthz
-					}
-				}
-				c.waitOnClusterReady()
-				checkStreams(t)
-			}
-		}
-		rollout(t)
-		checkStreams(t)
-	}
-	wg.Wait()
-	checkStreams(t)
 }
 
 func TestJetStreamClusterKeyValueLastSeqMismatch(t *testing.T) {
